% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_core_question_select.R
\name{mod_core_question_select_ui}
\alias{mod_core_question_select_ui}
\alias{mod_core_question_select_server}
\title{mod_core_question_select_ui and mod_core_question_select_server}
\usage{
mod_core_question_select_ui(id)

mod_core_question_select_server(input, output, session)
}
\arguments{
\item{id}{shiny id}

\item{input}{internal}

\item{output}{internal}

\item{session}{internal}
}
\description{
A shiny Module.
}
\keyword{internal}
