% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_test_text.R
\name{mod_test_text_ui}
\alias{mod_test_text_ui}
\alias{mod_test_text_server}
\title{mod_test_text_ui and mod_test_text_server}
\usage{
mod_test_text_ui(id)

mod_test_text_server(input, output, session, question)
}
\arguments{
\item{id}{shiny id}

\item{input}{internal}

\item{output}{internal}

\item{session}{internal}
}
\description{
A shiny Module.
}
\keyword{internal}
